+++ packages/terra-form-textarea/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Added
+  * Prop `onInput` - Callback for when user inputs a value.
+
 ## 5.3.0 - (September 1, 2020)
 
 * Removed